LayoutTests/fast/files/blob-text-gc-expected.txt

Make sure the Blob.text() promise gets resolved even if the Blob object is not kept alive by JS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text is "foobar0"
PASS text is "foobar1"
PASS text is "foobar2"
PASS text is "foobar3"
PASS text is "foobar4"
PASS text is "foobar5"
PASS text is "foobar6"
PASS text is "foobar7"
PASS text is "foobar8"
PASS text is "foobar9"
PASS text is "foobar10"
PASS text is "foobar11"
PASS text is "foobar12"
PASS text is "foobar13"
PASS text is "foobar14"
PASS text is "foobar15"
PASS text is "foobar16"
PASS text is "foobar17"
PASS text is "foobar18"
PASS text is "foobar19"
PASS text is "foobar20"
PASS text is "foobar21"
PASS text is "foobar22"
PASS text is "foobar23"
PASS text is "foobar24"
PASS text is "foobar25"
PASS text is "foobar26"
PASS text is "foobar27"
PASS text is "foobar28"
PASS text is "foobar29"
PASS text is "foobar30"
PASS text is "foobar31"
PASS text is "foobar32"
PASS text is "foobar33"
PASS text is "foobar34"
PASS text is "foobar35"
PASS text is "foobar36"
PASS text is "foobar37"
PASS text is "foobar38"
PASS text is "foobar39"
PASS text is "foobar40"
PASS text is "foobar41"
PASS text is "foobar42"
PASS text is "foobar43"
PASS text is "foobar44"
PASS text is "foobar45"
PASS text is "foobar46"
PASS text is "foobar47"
PASS text is "foobar48"
PASS text is "foobar49"
PASS text is "foobar50"
PASS text is "foobar51"
PASS text is "foobar52"
PASS text is "foobar53"
PASS text is "foobar54"
PASS text is "foobar55"
PASS text is "foobar56"
PASS text is "foobar57"
PASS text is "foobar58"
PASS text is "foobar59"
PASS text is "foobar60"
PASS text is "foobar61"
PASS text is "foobar62"
PASS text is "foobar63"
PASS text is "foobar64"
PASS text is "foobar65"
PASS text is "foobar66"
PASS text is "foobar67"
PASS text is "foobar68"
PASS text is "foobar69"
PASS text is "foobar70"
PASS text is "foobar71"
PASS text is "foobar72"
PASS text is "foobar73"
PASS text is "foobar74"
PASS text is "foobar75"
PASS text is "foobar76"
PASS text is "foobar77"
PASS text is "foobar78"
PASS text is "foobar79"
PASS text is "foobar80"
PASS text is "foobar81"
PASS text is "foobar82"
PASS text is "foobar83"
PASS text is "foobar84"
PASS text is "foobar85"
PASS text is "foobar86"
PASS text is "foobar87"
PASS text is "foobar88"
PASS text is "foobar89"
PASS text is "foobar90"
PASS text is "foobar91"
PASS text is "foobar92"
PASS text is "foobar93"
PASS text is "foobar94"
PASS text is "foobar95"
PASS text is "foobar96"
PASS text is "foobar97"
PASS text is "foobar98"
PASS text is "foobar99"
PASS successfullyParsed is true

TEST COMPLETE
